refactor(scheduler): address DRA extended resources review feedback
- Move shared/fractional GPU rejection from isTaskAllocatableOnNonAllocatedResources
  to PredicateByNodeResourcesType (scoped to IsSharedGPURequest), fixing false
  rejection of CPU-only pods on DRA-only nodes
- Unify ExtendedResourceClaimAllocation + ExtendedResourceClaimUID into a single
  *ExtendedResourceClaim{UID, Allocation} on PodInfo
- Drop PodsIndex+1 loop-start and i>PodsIndex guard in favour of simpler
  for i := range len(...) with an allocatable==0 check
- Drop ni.DeviceClassByResource != nil nil-guard (GetDeviceClass is nil-safe)
- Build DeviceClassByResource before snapshotNodes so each NodeInfo gets the cache
  at construction time instead of a second pass
- Remove stale-claim removal in preFilter (RemoveClaimPendingAllocation is a no-op
  on API-persisted claims; the binder owns that cleanup)
- Move ported upstream functions (createRequestsAndMappings and helpers) into a
  dedicated extended_resources_upstream.go so the file can be replaced wholesale
  on upstream upgrades
